requirements.txt
pynput
keyboard
pystray
Pillow
psutil